requirements.txt
jira
atlassian-python-api  
streamlit